cardioauth/vector_store/client.py
"""Vector store client for payer policy retrieval."""

from __future__ import annotations

import logging
from typing import Any

from cardioauth.config import Config

logger = logging.getLogger(__name__)


class VectorStoreClient:
    """Abstraction over Pinecone for policy document retrieval.

    In production, this would embed the query and search the Pinecone index.
    For now it provides the interface the PolicyAgent depends on.
    """

    def __init__(self, config: Config) -> None:
        self.config = config
        self._index = None

    def _get_index(self) -> Any:
        if self._index is None:
            from pinecone import Pinecone

            pc = Pinecone(api_key=self.config.pinecone_api_key)
            self._index = pc.Index(self.config.pinecone_index)
        return self._index

    def search(self, query: str, top_k: int = 5) -> list[dict[str, Any]]:
        """Search for relevant policy documents."""
        logger.info("VectorStore: searching for '%s' (top_k=%d)", query[:80], top_k)

        try:
            index = self._get_index()
            # In production: embed query with same model used for indexing,
            # then query Pinecone. For now, return empty to allow development.
            # results = index.query(vector=embedded_query, top_k=top_k, include_metadata=True)
            # return [match.metadata for match in results.matches]
            logger.warning("VectorStore: returning empty results — implement embedding + query")
            return []
        except Exception as e:
            logger.error("VectorStore: search failed: %s", e)
            return []

    def ingest_learning(self, learning_payload: dict[str, Any]) -> None:
        """Ingest outcome data back into the policy knowledge base."""
        logger.info("VectorStore: ingesting learning payload for %s / %s",
                     learning_payload.get("payer"), learning_payload.get("procedure"))
        # In production: embed and upsert into Pinecone
