requirements.txt
anthropic>=0.40.0
pydantic>=2.0
requests>=2.31
pinecone-client>=3.0
